Move tests to mpt circuit from mpt update
