+++ src/gadgets/mpt_update.rs
@@ -2057,472 +2057,3 @@ pub fn mpt_update_keys(proofs: &[Proof]) -> Vec<Fr> {
     }
     keys
 }
-
-#[cfg(test)]
-mod test {
-    use super::super::{
-        byte_bit::ByteBitGadget, byte_representation::ByteRepresentationConfig,
-        canonical_representation::CanonicalRepresentationConfig, key_bit::KeyBitConfig,
-        poseidon::PoseidonTable,
-    };
-    use super::*;
-    use crate::{constraint_builder::SelectorColumn, serde::SMTTrace, util::u256_hi_lo};
-    use halo2_proofs::{
-        circuit::{Layouter, SimpleFloorPlanner},
-        dev::MockProver,
-        halo2curves::bn256::Fr,
-        plonk::{Circuit, Error},
-    };
-    //use hash_circuit::hash::{PoseidonHashChip, PoseidonHashConfig, PoseidonHashTable};
-
-    const HASH_BLOCK_STEP_SIZE: usize = 62;
-    const N_ROWS: usize = 1024;
-
-    lazy_static! {
-        static ref EMPTY_STORAGE_PROOF_TYPE_2: (MPTProofType, SMTTrace) = (
-            MPTProofType::StorageDoesNotExist,
-            serde_json::from_str(include_str!(
-                "../../tests/generated/storage/empty_storage_proof_type_2.json"
-            ))
-            .unwrap(),
-        );
-        static ref NONCE_WRITE_TYPE_2_EMPTY_ACCOUNT: (MPTProofType, SMTTrace) = (
-            MPTProofType::NonceChanged,
-            serde_json::from_str(include_str!(
-                "../../tests/dual_code_hash/nonce_write_type_2_empty_account.json"
-            ))
-            .unwrap(),
-        );
-        static ref EMPTY_ACCOUNT_PROOF_TYPE_2: (MPTProofType, SMTTrace) = (
-            MPTProofType::AccountDoesNotExist,
-            serde_json::from_str(include_str!(
-                "../../tests/dual_code_hash/type_2_empty_account.json"
-            ))
-            .unwrap(),
-        );
-        static ref EMPTY_STORAGE_PROOF_SINGLETON_TRIE: (MPTProofType, SMTTrace) = (
-            MPTProofType::StorageDoesNotExist,
-            serde_json::from_str(include_str!(
-                "../../tests/generated/storage/empty_storage_proof_singleton_trie.json"
-            ))
-            .unwrap(),
-        );
-    }
-
-    #[derive(Clone, Debug, Default)]
-    struct TestCircuit {
-        n_rows: usize,
-        proofs: Vec<Proof>,
-    }
-
-    impl TestCircuit {
-        fn new(n_rows: usize, traces: Vec<(MPTProofType, SMTTrace)>) -> Self {
-            Self {
-                n_rows,
-                proofs: traces.into_iter().map(Proof::from).collect(),
-            }
-        }
-
-        fn hash_traces(&self) -> Vec<(Fr, Fr, Fr)> {
-            hash_traces(&self.proofs)
-        }
-        fn byte_representations(&self) -> (Vec<u64>, Vec<u128>, Vec<Fr>) {
-            byte_representations(&self.proofs)
-        }
-        fn key_bit_lookups(&self) -> Vec<(Fr, usize, bool)> {
-            key_bit_lookups(&self.proofs)
-        }
-        fn keys(&self) -> Vec<Fr> {
-            mpt_update_keys(&self.proofs)
-        }
-    }
-
-    impl Circuit<Fr> for TestCircuit {
-        type Config = (
-            SelectorColumn,
-            MptUpdateConfig,
-            PoseidonTable,
-            CanonicalRepresentationConfig,
-            KeyBitConfig,
-            ByteBitGadget,
-            ByteRepresentationConfig,
-            RlcRandomness,
-            //PoseidonHashConfig<Fr>,
-        );
-        type FloorPlanner = SimpleFloorPlanner;
-
-        fn without_witnesses(&self) -> Self {
-            Self::default()
-        }
-
-        fn configure(cs: &mut ConstraintSystem<Fr>) -> Self::Config {
-            let selector = SelectorColumn(cs.fixed_column());
-            let mut cb = ConstraintBuilder::new(selector);
-            let rlc_randomness = RlcRandomness::configure(cs);
-
-            let poseidon = PoseidonTable::dev_configure(cs, &mut cb);
-            let byte_bit = ByteBitGadget::configure(cs, &mut cb);
-            let byte_representation =
-                ByteRepresentationConfig::configure(cs, &mut cb, &byte_bit, &rlc_randomness);
-            let canonical_representation =
-                CanonicalRepresentationConfig::configure(cs, &mut cb, &byte_bit);
-            let key_bit = KeyBitConfig::configure(
-                cs,
-                &mut cb,
-                &canonical_representation,
-                &byte_bit,
-                &byte_bit,
-                &byte_bit,
-            );
-
-            let mpt_update = MptUpdateConfig::configure(
-                cs,
-                &mut cb,
-                &poseidon,
-                &key_bit,
-                &byte_representation,
-                &byte_representation,
-                &rlc_randomness,
-            );
-
-            cb.build(cs);
-
-            //            let poseidon_config =
-            //                PoseidonHashConfig::configure_sub(cs, poseidon.columns(), HASH_BLOCK_STEP_SIZE);
-
-            (
-                selector,
-                mpt_update,
-                poseidon,
-                canonical_representation,
-                key_bit,
-                byte_bit,
-                byte_representation,
-                rlc_randomness,
-                //                poseidon_config,
-            )
-        }
-
-        fn synthesize(
-            &self,
-            config: Self::Config,
-            mut layouter: impl Layouter<Fr>,
-        ) -> Result<(), Error> {
-            let (
-                selector,
-                mpt_update,
-                poseidon,
-                canonical_representation,
-                key_bit,
-                byte_bit,
-                byte_representation,
-                rlc_randomness,
-                //                poseidon_config,
-            ) = config;
-
-            let (u64s, u128s, frs) = self.byte_representations();
-            let randomness = rlc_randomness.value(&layouter);
-
-            /*            let mut poseidon_hash_table = PoseidonHashTable::default();
-                        poseidon_hash_table.constant_inputs_with_check(&self.hash_traces());
-                        let poseidon_chip = PoseidonHashChip::<_, HASH_BLOCK_STEP_SIZE>::construct(
-                            poseidon_config,
-                            &poseidon_hash_table,
-                            5000,
-                            false,
-                            Some(Fr::one()),
-                        );
-            */
-            layouter.assign_region(
-                || "hash table",
-                |mut region| {
-                    for offset in 0..self.n_rows {
-                        selector.enable(&mut region, offset);
-                    }
-                    let n_mpt_rows = mpt_update.assign(&mut region, &self.proofs, randomness);
-                    assert!(n_mpt_rows <= self.n_rows, "{:?}", n_mpt_rows);
-                    for offset in n_mpt_rows..self.n_rows {
-                        mpt_update.assign_padding_row(&mut region, offset);
-                    }
-                    poseidon.dev_load(&mut region, &self.hash_traces(), 5000);
-                    canonical_representation.assign(&mut region, &self.keys());
-                    key_bit.assign(&mut region, &self.key_bit_lookups());
-                    byte_bit.assign(&mut region);
-                    byte_representation.assign(&mut region, &u64s, &u128s, &frs, randomness);
-
-                    Ok(())
-                },
-            )?;
-
-            // poseidon_chip.load(&mut layouter)?;
-
-            Ok(())
-        }
-    }
-
-    fn mock_prove(proof_type: MPTProofType, trace: &str) {
-        let circuit = TestCircuit::new(
-            N_ROWS,
-            vec![(proof_type, serde_json::from_str(trace).unwrap())],
-        );
-        let prover = MockProver::<Fr>::run(14, &circuit, vec![]).unwrap();
-        assert_eq!(prover.verify(), Ok(()));
-    }
-
-    #[test]
-    fn test_empty() {
-        let circuit = TestCircuit {
-            n_rows: N_ROWS,
-            proofs: vec![],
-        };
-        let prover = MockProver::<Fr>::run(14, &circuit, vec![]).unwrap();
-        assert_eq!(prover.verify(), Ok(()));
-    }
-
-    #[test]
-    fn nonce_write_existing_account() {
-        mock_prove(
-            MPTProofType::NonceChanged,
-            include_str!("../../tests/dual_code_hash/nonce_write_existing_account.json"),
-        );
-    }
-
-    #[test]
-    fn nonce_write_type_1_empty_account() {
-        mock_prove(
-            MPTProofType::NonceChanged,
-            include_str!("../../tests/dual_code_hash/nonce_write_type_1_empty_account.json"),
-        );
-    }
-
-    #[test]
-    fn nonce_write_type_2_empty_account() {
-        mock_prove(
-            MPTProofType::NonceChanged,
-            include_str!("../../tests/dual_code_hash/nonce_write_type_2_empty_account.json"),
-        );
-    }
-
-    #[test]
-    fn nonce_update_existing() {
-        mock_prove(
-            MPTProofType::NonceChanged,
-            include_str!("../../tests/generated/nonce_update_existing.json"),
-        );
-    }
-
-    #[test]
-    fn nonce_update_type_1() {
-        mock_prove(
-            MPTProofType::NonceChanged,
-            include_str!("../../tests/generated/nonce_update_type_1.json"),
-        );
-    }
-
-    #[test]
-    fn nonce_update_type_2() {
-        mock_prove(
-            MPTProofType::NonceChanged,
-            include_str!("../../tests/generated/nonce_update_type_2.json"),
-        );
-    }
-
-    #[test]
-    fn balance_update_existing() {
-        mock_prove(
-            MPTProofType::BalanceChanged,
-            include_str!("../../tests/generated/balance_update_existing.json"),
-        );
-    }
-
-    #[test]
-    fn balance_update_type_1() {
-        mock_prove(
-            MPTProofType::BalanceChanged,
-            include_str!("../../tests/generated/balance_update_type_1.json"),
-        );
-    }
-
-    #[test]
-    fn balance_update_type_2() {
-        mock_prove(
-            MPTProofType::BalanceChanged,
-            include_str!("../../tests/generated/balance_update_type_2.json"),
-        );
-    }
-
-    #[test]
-    fn code_size_update_existing() {
-        mock_prove(
-            MPTProofType::CodeSizeExists,
-            include_str!("../../tests/generated/code_size_update_existing.json"),
-        );
-    }
-
-    #[test]
-    fn code_hash_update_existing() {
-        mock_prove(
-            MPTProofType::PoseidonCodeHashExists,
-            include_str!("../../tests/generated/code_hash_update_existing.json"),
-        );
-    }
-
-    #[test]
-    fn keccak_code_hash_update_existing() {
-        mock_prove(
-            MPTProofType::CodeHashExists, // yes, the naming is very confusing :/
-            include_str!("../../tests/generated/keccak_code_hash_update_existing.json"),
-        );
-    }
-
-    #[test]
-    fn keccak_code_hash_read_existing() {
-        mock_prove(
-            MPTProofType::CodeHashExists, // yes, the naming is very confusing :/
-            include_str!("../../tests/generated/keccak_code_hash_read_existing.json"),
-        );
-    }
-
-    #[test]
-    fn update_storage_existing_to_existing() {
-        mock_prove(
-            MPTProofType::StorageChanged,
-            include_str!("../../tests/generated/storage/update_storage_existing_to_existing.json"),
-        );
-    }
-
-    #[test]
-    fn nonexisting_type_1() {
-        mock_prove(
-            MPTProofType::AccountDoesNotExist,
-            include_str!("../../tests/dual_code_hash/type_1_empty_account.json"),
-        );
-    }
-
-    #[test]
-    fn write_empty_storage_trie() {
-        mock_prove(
-            MPTProofType::StorageChanged,
-            include_str!("../../tests/generated/storage/write_empty_storage_trie.json"),
-        );
-    }
-
-    #[test]
-    fn nonexisting_type_2() {
-        mock_prove(
-            MPTProofType::AccountDoesNotExist,
-            include_str!("../../tests/dual_code_hash/type_2_empty_account.json"),
-        );
-    }
-
-    #[test]
-    fn write_singleton_storage_trie() {
-        mock_prove(
-            MPTProofType::StorageChanged,
-            include_str!("../../tests/generated/storage/write_singleton_storage_trie.json"),
-        );
-    }
-
-    #[test]
-    fn write_zero_singleton_storage_trie() {
-        mock_prove(
-            MPTProofType::StorageChanged,
-            include_str!("../../tests/generated/storage/write_zero_singleton_storage_trie.json"),
-        );
-    }
-
-    #[test]
-    fn write_zero_doubleton_storage_trie() {
-        mock_prove(
-            MPTProofType::StorageChanged,
-            include_str!("../../tests/generated/storage/write_zero_doubleton_storage_trie.json"),
-        );
-    }
-
-    #[test]
-    fn write_zero_storage_trie() {
-        mock_prove(
-            MPTProofType::StorageChanged,
-            include_str!("../../tests/generated/storage/write_zero_storage_trie.json"),
-        );
-    }
-
-    #[test]
-    fn empty_storage_proof_empty_trie() {
-        mock_prove(
-            MPTProofType::StorageDoesNotExist,
-            include_str!("../../tests/generated/storage/empty_storage_proof_empty_trie.json"),
-        );
-    }
-
-    #[test]
-    fn empty_storage_proof_singleton_trie() {
-        mock_prove(
-            MPTProofType::StorageDoesNotExist,
-            include_str!("../../tests/generated/storage/empty_storage_proof_singleton_trie.json"),
-        );
-    }
-
-    #[test]
-    fn empty_storage_proof_type_1() {
-        mock_prove(
-            MPTProofType::StorageDoesNotExist,
-            include_str!("../../tests/generated/storage/empty_storage_proof_type_1.json"),
-        );
-    }
-
-    #[test]
-    fn empty_storage_proof_type_2() {
-        mock_prove(
-            MPTProofType::StorageDoesNotExist,
-            include_str!("../../tests/generated/storage/empty_storage_proof_type_2.json"),
-        );
-    }
-
-    #[test]
-    fn test_account_key() {
-        for address in vec![Address::zero(), Address::repeat_byte(0x56)] {
-            assert_eq!(
-                hash(
-                    Fr::from_u128(address_high(address)),
-                    Fr::from_u128(address_low(address)),
-                ),
-                account_key(address)
-            );
-        }
-    }
-
-    #[test]
-    fn prove_updates() {
-        let updates = vec![
-            EMPTY_STORAGE_PROOF_TYPE_2.clone(),
-            EMPTY_STORAGE_PROOF_SINGLETON_TRIE.clone(),
-            EMPTY_ACCOUNT_PROOF_TYPE_2.clone(),
-            NONCE_WRITE_TYPE_2_EMPTY_ACCOUNT.clone(),
-        ];
-
-        let circuit = TestCircuit::new(N_ROWS, updates);
-        let prover = MockProver::<Fr>::run(14, &circuit, vec![]).unwrap();
-        assert_eq!(prover.verify(), Ok(()));
-    }
-
-    #[test]
-    fn empty_account_empty_storage() {
-        mock_prove(
-            MPTProofType::StorageDoesNotExist,
-            include_str!("../../tests/generated/storage/empty_account_empty_storage_proof.json"),
-        );
-    }
-    /*
-    #[test]
-    fn uniswapv2_factory_create_pair() {
-        let updates = serde_json::from_str(include_str!(
-            "../../tests/mock_test_traces/uniswapv2_factory-createPair.json"
-        ))
-        .unwrap();
-        let circuit = TestCircuit::new(4096, updates);
-        let prover = MockProver::<Fr>::run(16, &circuit, vec![]).unwrap();
-        assert_eq!(prover.verify(), Ok(()));
-    }
-    */
-}
